internal/cli: migrated the cli to using new budfs. framework/generator: replaced old generator generator.
