+++ framework/generator/generator.go
@@ -3,18 +3,15 @@ package generator
 import (
 	_ "embed"
 	"fmt"
+	"io/fs"
 
-	"github.com/livebud/bud/framework"
-	"github.com/livebud/bud/internal/gobuild"
-
-	"github.com/livebud/bud/package/di"
+	"github.com/livebud/bud/internal/bail"
+	"github.com/livebud/bud/internal/gotemplate"
+	"github.com/livebud/bud/internal/imports"
+	"github.com/livebud/bud/package/genfs"
+	"github.com/livebud/bud/package/gomod"
 	"github.com/livebud/bud/package/log"
 	"github.com/livebud/bud/package/parser"
-	"github.com/livebud/bud/package/remotefs"
-
-	"github.com/livebud/bud/internal/gotemplate"
-	"github.com/livebud/bud/package/budfs"
-	"github.com/livebud/bud/package/gomod"
 )
 
 //go:embed generator.gotext
@@ -26,26 +23,19 @@ func Generate(state *State) ([]byte, error) {
 	return generator.Generate(state)
 }
 
-func New(bfs *budfs.FileSystem, flag *framework.Flag, injector *di.Injector, log log.Log, module *gomod.Module, parser *parser.Parser) *Generator {
-	return &Generator{bfs, flag, injector, log, module, parser, nil}
+func New(log log.Log, module *gomod.Module, parser *parser.Parser) *Generator {
+	return &Generator{log, module, parser}
 }
 
 type Generator struct {
-	bfs      *budfs.FileSystem
-	flag     *framework.Flag
-	injector *di.Injector
-	log      log.Log
-	module   *gomod.Module
-	parser   *parser.Parser
-
-	// process starts as nil
-	process *remotefs.Process
+	log    log.Log
+	module *gomod.Module
+	parser *parser.Parser
 }
 
 // GenerateFile connects to the remotefs and mounts the remote directory.
-func (g *Generator) GenerateFileOld(fsys budfs.FS, file *budfs.File) error {
-	g.log.Debug("framework/generator: generating the main.go service containing the generators")
-	state, err := Load(fsys, g.injector, g.log, g.module, g.parser)
+func (g *Generator) GenerateFile(fsys genfs.FS, file *genfs.File) error {
+	state, err := g.Load(fsys)
 	if err != nil {
 		return fmt.Errorf("framework/generator: unable to load. %w", err)
 	}
@@ -54,56 +44,11 @@ func (g *Generator) GenerateFileOld(fsys budfs.FS, file *budfs.File) error {
 		return err
 	}
 	file.Data = code
-
-	g.log.Debug("framework/generator: write the generator main.go file to bud/command/.generate/main.go")
-	if err := g.module.MkdirAll("bud/command/.generate", 0755); err != nil {
-		return err
-	}
-	if err := g.module.WriteFile("bud/command/.generate/main.go", []byte(code), 0644); err != nil {
-		return err
-	}
-
-	g.log.Debug("framework/generator: build the main.go file to bud/.generate")
-	builder := gobuild.New(g.module)
-	builder.Env = append([]string{}, g.flag.Env...)
-	builder.Stderr = g.flag.Stderr
-	builder.Stdout = g.flag.Stdout
-	if err := builder.Build(fsys.Context(), "bud/command/.generate/main.go", "bud/.generate"); err != nil {
-		return fmt.Errorf("framework/generator: unable to build 'bud/.generate'. %s", err)
-	}
-
-	if g.process != nil {
-		g.log.Debug("framework/generator: closing existing process")
-		if err := g.process.Close(); err != nil {
-			return err
-		}
-		g.process = nil
-	}
-
-	g.log.Debug("framework/generator: start bud/.generate that will serve the remote filesystem")
-	cmd := &remotefs.Command{
-		Dir:    g.module.Directory(),
-		Env:    append([]string{}, g.flag.Env...),
-		Stderr: g.flag.Stderr,
-		Stdout: g.flag.Stdout,
-	}
-	g.process, err = cmd.Start(fsys.Context(), g.module.Directory("bud/.generate"))
-	if err != nil {
-		return err
-	}
-
-	// Close the process when the filesystem is closed.
-	fsys.Defer(func() error {
-		if g.process == nil {
-			return nil
-		}
-		g.log.Debug("framework/generator: shutting down the remotefs")
-		return g.process.Close()
-	})
-
-	// Mount the remote filesystem
-	g.log.Debug("framework/generator: mounting the running remote filesystem")
-	g.bfs.Mount(g.process)
-
 	return nil
+}
+
+// Load the generators
+func (g *Generator) Load(fsys fs.FS) (*State, error) {
+	loader := &loader{g.log, g.module, g.parser, imports.New(), bail.Struct{}}
+	return loader.Load(fsys)
 }
