--- framework/generator2/generator.gotext
@@ -1,26 +1,0 @@
-package generator
-
-{{- if $.Imports }}
-
-import (
-	{{- range $import := $.Imports }}
-	{{$import.Name}} "{{$import.Path}}"
-	{{- end }}
-)
-{{- end }}
-
-func New(
-	fsys *genfs.FileSystem,
-	module *gomod.Module,
-	log log.Log,
-	{{- range $generator := $.Generators }}
-	{{ $generator.Camel }} *{{ $generator.Import.Name }}.Generator,
-	{{- end }}
-) FS {
-	{{- range $generator := $.Generators }}
-	fsys.{{ $generator.Type }}(`{{ $generator.Path }}`, {{ $generator.Camel }})
-	{{- end }}
-	return fsys
-}
-
-type FS = fs.FS
